Refactor formatTime()
