+++ interfaces/console/utils/formatTime.ts
@@ -1,0 +1,9 @@
+export const formatTime = (isoString: string) => {
+    const date = new Date(isoString);
+    const day = date.getDate().toString().padStart(2, '0');
+    const month = (date.getMonth() + 1).toString();
+    const hours = date.getHours();
+    const period = hours >= 12 ? 'PM' : 'AM';
+    const formattedHours = (hours % 12 || 12).toString();
+    return `${month}/${day} ${formattedHours}${period}`;
+  };
